requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
limits==5.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyreadline3==3.5.4
python-dotenv==1.1.1
python-multipart==0.0.20
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wrapt==1.17.3
